requirements_old.txt
# Django and core dependencies
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.2
python-dotenv>=1.0.0

# Authentication and Security
# djangorestframework-authtoken is included in djangorestframework

# WebSocket support
channels>=4.0.0
channels-redis>=4.0.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter
redis>=4.5.0

# File handling and storage
Pillow>=10.0.0  # Image processing
boto3>=1.26.0  # AWS S3 storage
django-storages>=1.13.0  # Storage backends

# PDF processing, AI APIs, and payment processing removed
# PyPDF2>=3.0.0  # PDF processing - REMOVED
# pdf2image>=1.16.0  # PDF to image conversion - REMOVED
# openai>=1.0.0  # OpenAI API - REMOVED
# anthropic>=0.3.0  # Anthropic API - REMOVED
# stripe>=5.0.0  # Payment processing - REMOVED

# Task queue
celery>=5.3.0
django-celery-beat>=2.5.0

# Email
django-anymail>=10.0

# Development tools
django-debug-toolbar>=4.0.0

# Production
gunicorn>=21.0.0
whitenoise>=6.5.0

# Testing
pytest>=7.0.0
pytest-django>=4.5.0

# Monitoring and logging
sentry-sdk>=1.0.0

# Other utilities
python-dateutil>=2.8.0
requests>=2.31.0
urllib3>=2.0.0